+++ core/src/uix/core.clj
@@ -73,7 +73,8 @@
   "Creates UIx component. Similar to defn, but doesn't support multi arity.
   A component should have a single argument of props."
   [sym & fdecl]
-  (let [[fname args fdecl] (parse-defui-sig `defui sym fdecl)]
+  (let [[fname args fdecl] (parse-sig `defui sym fdecl)]
+    (uix.linter/lint! sym fdecl &form &env)
     (if (uix.lib/cljs-env? &env)
       (let [var-sym (-> (str (-> &env :ns :name) "/" sym) symbol (with-meta {:tag 'js}))
             body (uix.dev/with-fast-refresh var-sym fdecl)]
@@ -95,7 +96,8 @@
   (let [[sym fdecl] (if (symbol? (first fdecl))
                       [(first fdecl) (rest fdecl)]
                       [(gensym "uix-fn") fdecl])
-        [fname args body] (parse-defui-sig `fn sym fdecl)]
+        [fname args body] (parse-sig `fn sym fdecl)]
+    (uix.linter/lint! sym body &form &env)
     (if (uix.lib/cljs-env? &env)
       (let [var-sym (with-meta sym {:tag 'js})]
         (uix.linter/lint! sym body &env)
@@ -119,8 +121,10 @@
   DOM element: ($ :button#id.class {:on-click handle-click} \"click me\")
   React component: ($ title-bar {:title \"Title\"})"
   ([tag]
+   (uix.linter/lint-element* &form &env)
    (uix.compiler.aot/compile-element [tag] {:env &env}))
   ([tag props & children]
+   (uix.linter/lint-element* &form &env)
    (uix.compiler.aot/compile-element (into [tag props] children) {:env &env})))
 
 (defn parse-defhook-sig [sym fdecl]
@@ -176,6 +180,17 @@
    (make-hook-with-deps 'uix.hooks.alpha/use-layout-effect &env &form f nil))
   ([f deps]
    (make-hook-with-deps 'uix.hooks.alpha/use-layout-effect &env &form f deps)))
+
+(defmacro use-insertion-effect
+  "Takes a function to be executed synchronously before all DOM mutations
+  and optional vector of dependencies. Use this to inject styles into the DOM
+  before reading layout in `useLayoutEffect`.
+
+  See: https://reactjs.org/docs/hooks-reference.html#useinsertioneffect"
+  ([f]
+   (make-hook-with-deps 'uix.hooks.alpha/use-insertion-effect &env &form f nil))
+  ([f deps]
+   (make-hook-with-deps 'uix.hooks.alpha/use-insertion-effect &env &form f deps)))
 
 (defmacro use-memo
   "Takes function f and required vector of dependencies, and returns memoized result of f.