+++ core/src/uix/linter.clj
@@ -49,30 +49,33 @@
   (contains? effect-hooks (name (first form))))
 
 (defn form->loc [form]
-  (select-keys form [:line :column]))
+  (select-keys (meta form) [:line :column]))
 
 (defn find-env-for-form [type form]
   (case type
     (::hook-in-branch ::hook-in-loop
                       ::deps-coll-literal ::literal-value-in-deps
                       ::unsafe-set-state ::missing-key ::non-defhook-hook)
-    (form->loc (meta form))
+    (form->loc form)
 
     ::inline-function
-    (form->loc (meta (second form)))
+    (form->loc (second form))
 
     ::deps-array-literal
-    (form->loc (meta (.-val form)))
+    (form->loc (.-val form))
 
     nil))
 
-(defn add-error! [form type]
-  (swap! *context* update :errors conj {:source form
-                                        :source-context *source-context*
-                                        :type type
-                                        :env (find-env-for-form type form)}))
+(defn add-error!
+  ([form type]
+   (add-error! form type (find-env-for-form type form)))
+  ([form type env]
+   (swap! *component-context* update :errors conj {:source form
+                                                   :source-context *source-context*
+                                                   :type type
+                                                   :env env})))
 
-(defn- uix-element? [form]
+(defn uix-element? [form]
   (and (list? form) (= '$ (first form))))
 
 (defn- missing-key? [[_ _ attrs :as form]]
@@ -287,10 +290,13 @@
   (read-re-frame-config))
 
 (defn lint-re-frame! [form env]
-  (let [sources (->> (uix.lib/find-form rf-subscribe-call? form)
-                     (keep #(let [v (ana/resolve-var env (first %))]
-                              (when (contains? re-frame-config (:name v))
-                                (assoc v :source %)))))]
+  (let [resolve-fn (if (uix.lib/cljs-env? env)
+                     ana/resolve-var
+                     resolve)
+        sources    (->> (uix.lib/find-form rf-subscribe-call? form)
+                        (keep #(let [v (resolve-fn env (first %))]
+                                 (when (contains? re-frame-config (:name v))
+                                   (assoc v :source %)))))]
     (run! #(ana/warning ::non-reactive-re-frame-subscribe env %)
           sources)))
 
@@ -299,19 +305,31 @@
        (:name v) ", use `use-subscribe` hook instead.\n"
        "Read https://github.com/pitch-io/uix/blob/master/docs/interop-with-reagent.md#syncing-with-ratoms-and-re-frame for more context"))
 
-(defn lint! [sym form env]
-  (binding [*context* (atom {:errors [] :env env})]
-    (lint-body! form)
-    (lint-re-frame! form env)
-    (let [{:keys [errors]} @*context*
-          {:keys [column line]} env]
-      (doseq [err errors]
-        (ana/warning (:type err)
-                     (or (:env err) env)
-                     (into {:name (str (-> env :ns :name) "/" sym)
-                            :column column
-                            :line line}
-                           err))))))
+(defmulti lint-component (fn [type form env]))
+(defmulti lint-element (fn [type form env]))
+(defmulti lint-hook-with-deps (fn [type form env]))
+
+(defn- run-linters! [mf & args]
+  (doseq [[key f] (.getMethodTable ^clojure.lang.MultiFn mf)]
+    (apply f key args)))
+
+(defn- report-errors!
+  ([env]
+   (report-errors! env nil))
+  ([env m]
+   (let [{:keys [errors]} @*component-context*
+         {:keys [column line]} env]
+     (run! #(ana/warning (:type %)
+                         (or (:env %) env)
+                         (merge {:column column :line line} m %))
+           errors))))
+
+(defn lint! [sym body form env]
+  (binding [*component-context* (atom {:errors []})]
+    (lint-body! body)
+    (lint-re-frame! body env)
+    (run-linters! lint-component form env)
+    (report-errors! env {:name (str (-> env :ns :name) "/" sym)})))
 
 ;; === Exhaustive Deps ===
 
@@ -341,7 +359,8 @@
                        (get-in env [:locals (:name %) :name]) ;; from an outer scope
                        (-> % :info :shadow not) ;; but not a local shadowing locals from outer scope
                        (not (deps (:name %))))) ;; and not declared in deps vector
-         (map :name))))
+         (map :name)
+         distinct)))
 
 (defmethod pp/code-dispatch JSValue [alis]
   (.write ^Writer *out* "#js ")
@@ -496,11 +515,10 @@
     ;; when a reference to a function passed into a hook, should be an inline function instead
     (not (fn-literal? f)) [::inline-function {:source form :env (find-env-for-form ::inline-function form)}]
 
-    (vector? deps)
+    (and (vector? deps) (not (:lint/disable (meta deps))))
     (let [[missing-deps declared-unnecessary-deps suggested-deps] (find-missing-and-unnecessary-deps env f deps)]
       ;; when hook function is referencing vars from out scope that are missing in deps vector
-      (when (or (and (:lint-deps (meta deps)) (seq missing-deps))
-                (seq declared-unnecessary-deps))
+      (when (or (seq missing-deps) (seq declared-unnecessary-deps))
         [::missing-deps {:missing-deps missing-deps
                          :unnecessary-deps declared-unnecessary-deps
                          :suggested-deps suggested-deps
@@ -520,5 +538,12 @@
 
 (defn lint-exhaustive-deps! [env form f deps]
   (doseq [[error-type opts] (lint-exhaustive-deps env form f deps)]
-    (ana/warning error-type (or (:env opts) env) opts)))
+    (ana/warning error-type (or (:env opts) env) opts))
+  (binding [*component-context* (atom {:errors []})]
+    (run-linters! lint-hook-with-deps form env)
+    (report-errors! env)))
 
+(defn lint-element* [form env]
+  (binding [*component-context* (atom {:errors []})]
+    (uix.linter/run-linters! uix.linter/lint-element form env)
+    (report-errors! env)))